requirements.txt
requests==2.22.0
beautifulsoup4==4.9.3
path.py==12.5.0
